TS: refactored PortNumberGenerator to be able to easier change IPv4-only impl. and busy adding reserve flag to next.
"reserve" will let the PortNumberGenerator keep the port-number(s) requested bound; to be released explicitly "just before using".

This to keep the time between unbound and in-use as short as possible to "prevent" another process/thread to claim it.  This can never really be enforced (the're ephemeral ports, not blacklisted) for testing without making lots more assumptions about the OS & Processes running there.